UI fixes (#23)
* fix tooltip triggering on trend items

AN-72 Hover on the arrow of the trend indicator does not show info tooltip

* fix trend cursor style

AN-73 Hover the trend indicator text should show default mouse cursor and not the text edit.

* fix totals tabs style

AN-71The tabs current item is not style correctly.

* fix trend arrow style

AN-74 The trend indicator should be bold.

* fix trend styles

* fix trend styles

* fix devices overview trend margin

AN-75 The trend indicator should be 12 px away from the metric on the right. Currently the indicator is aligned along a seperate column

* fix date filter styles

AN-77 In the period menu, the V of the selected item should be 16px away from the menu edges.

* fix graphs text font family

AN-80 All graphs tooltips arent using Lato font

* fix graph tooltip style

AN-84 All the graph tooltips padding should be 16px.

* fix toast message style

AN-86 The toast message after export should have padding top 16px

* fix totals tab paddings

AN-87 Tabs section should have padding right/left of 24px

* fix geo table col style

AN-89 In Geo, in the country drilldown the table display region/states name in the wrong style

* fix dropdown style

AN-93 Padding in the dropdowns and the menu is not respected

* fix table no results message

AN-95 While the module are loaded, No result message should not appear

* fix end users totals

AN-100 In the end user storage report, the Total Minutes are express in a timecode. It should a rounded number with the unit "Min"

* fix table's sort bar styles

AN-76 Sort bar height is incorrect. Not align with the general style.

* fix user filter placeholder